+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/choosers/SmartChooser.kt
@@ -3,20 +3,15 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.choosers
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.database.tables.PostsTable
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.commonLogger
 import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.rating.database.PostIdRatingPair
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.CalculatedDateTime
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.CalculatedPeriod
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.asPairs
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.parseDateTimes
 import com.github.insanusmokrassar.IObjectK.interfaces.IObject
 import com.github.insanusmokrassar.IObjectKRealisations.toObject
-import kotlinx.coroutines.experimental.async
-import kotlinx.coroutines.experimental.launch
 import org.joda.time.DateTime
 import org.joda.time.DateTimeZone
-import org.joda.time.format.DateTimeFormat
-import org.joda.time.format.DateTimeFormatter
 import java.util.*
-
-private val defaultTimeZone = DateTimeZone.getDefault()
-private val timeFormat: DateTimeFormatter = DateTimeFormat.forPattern("HH:mm")
-
-private val coverFormat: DateTimeFormatter = DateTimeFormat.forPattern("HH:mm:ss:SS")
 
 private const val ascendSort = "ascend"
 private const val descendSort = "descend"
@@ -82,28 +77,13 @@ private val commonInnerChoosers = mapOf<String, InnerChooser>(
 private class SmartChooserConfigItem (
     val minRate: Int? = null,
     val maxRate: Int? = null,
-    val timeOffset: String = defaultTimeZone.id,
-    val time: Array<String?> = arrayOf(
-        "00:00",
-        null
-    ),
+    val time: String? = null,
+    val times: List<String>? = null,
     val sort: String = defaultSort,
     val count: Int = 1,
     val minAge: Long? = null,
     val maxAge: Long? = null
 ) {
-    private val zeroHour: DateTime by lazy {
-        timeFormat.parseDateTime("00:00")
-    }
-
-    private val nextDayZeroHour: DateTime by lazy {
-        zeroHour.plusDays(1)
-    }
-
-    private val timeZone by lazy {
-        DateTimeZone.forID(timeOffset)
-    }
-
     private val minAgeAsDateTime: DateTime? by lazy {
         minAge ?.let {
             DateTime.now().withZone(DateTimeZone.UTC).withMillis(it).withZone(DateTimeZone.getDefault())
@@ -116,53 +96,20 @@ private class SmartChooserConfigItem (
         }
     }
 
-    private val timePairs: List<Pair<DateTime, DateTime>> by lazy {
-        val pairs = mutableListOf<Pair<DateTime, DateTime>>()
-        var currentPair: Pair<DateTime?, DateTime?>? = null
-
-        val timeFormatWithConfigTimeZone = timeFormat.withZone(timeZone)
-
-        time.forEach {
-            s ->
-            val dateTime = s ?.let {
-                timeFormatWithConfigTimeZone.parseDateTime(
-                    it
-                ).withZone(
-                    DateTimeZone.getDefault()
-                )
-            } ?.let {
-                it.withDate(
-                    zeroHour.toLocalDate()
-                )
-            }
-            currentPair ?.let {
-                (from, _) ->
-                val first = from ?: zeroHour
-                val second = dateTime ?: nextDayZeroHour
-
-                if (first > second) {
-                    pairs.add(first to nextDayZeroHour)
-                    pairs.add(zeroHour to second)
-                } else {
-                    pairs.add(first to second)
-                }
-
-                currentPair = null
-            } ?:let {
-                currentPair = dateTime to null
-            }
-        }
-        pairs
+    private val timePairs: List<CalculatedPeriod> by lazy {
+        (times ?.flatMap {
+            it.parseDateTimes()
+        } ?: time ?.parseDateTimes() ?: emptyList()).asPairs()
     }
 
     fun actual(
-        now: Long = timeFormat.parseMillis(
-            timeFormat.print(DateTime.now())
-        )
+        now: Long = System.currentTimeMillis()
     ): Boolean {
         timePairs.forEach {
             (from, to) ->
-            if ((from.isBefore(now) || from.isEqual(now)) && to.isAfter(now)) {
+            val fromAsFuture = from.asFuture
+            val toAsFuture = to.asFuture
+            if ((fromAsFuture.isBefore(now) || fromAsFuture.isEqual(now)) && toAsFuture.isAfter(now)) {
                 return true
             }
         }
@@ -191,32 +138,10 @@ private class SmartChooserConfigItem (
         stringBuilder.append("Time:\n")
         timePairs.forEach {
             (from, to) ->
-            stringBuilder.append("  ${timeFormat.print(from)} - ${timeFormat.print(to)}\n")
+            stringBuilder.append("  ${from.dateTime} - ${to.dateTime}\n")
         }
         return stringBuilder.toString()
     }
-}
-
-private fun checkCover(times: List<SmartChooserConfigItem>): List<Pair<Long, Long>> {
-    var currentFirst: Long? = null
-    val result = mutableListOf<Pair<Long, Long>>()
-    (timeFormat.parseMillis("00:00") .. timeFormat.parseDateTime("00:00").plusDays(1).millis).forEach {
-        now ->
-        currentFirst ?.let {
-            first ->
-            times.firstOrNull { it.actual(now) } ?.let {
-                if (first != now) {
-                    result.add(first to now)
-                }
-                currentFirst = null
-            }
-        } ?: {
-            times.firstOrNull { it.actual(now) } ?: {
-                currentFirst = now
-            }()
-        }()
-    }
-    return result
 }
 
 private class SmartChooserConfig(
@@ -229,25 +154,7 @@ class SmartChooser(
     private val config = config.toObject(SmartChooserConfig::class.java)
 
     init {
-        println("Smart chooser inited: ${this.config.times.joinToString(separator = "\n") { it.toString() }}")
-
-        launch {
-            val deferred = async {
-                checkCover(this@SmartChooser.config.times)
-            }
-
-            deferred.await().also {
-                if (it.isNotEmpty()) {
-                    it.joinToString("\n", "$name: Uncovered time:\n") {
-                        "${coverFormat.print(it.first)} - ${coverFormat.print(it.second)}"
-                    }.also {
-                        commonLogger.warning(it)
-                    }
-                } else {
-                    commonLogger.info("$name: All day covered")
-                }
-            }
-        }
+        commonLogger.info("Smart chooser inited: ${this.config.times.joinToString(separator = "\n") { it.toString() }}")
     }
 
     override fun triggerChoose(): Collection<Int> {